cleaned up stars
